backend/requirements.txt
Django~=5.1.6
PyYAML~=6.0.2
requests~=2.32.3
djangorestframework~=3.15.2
django-cors-headers==4.7.0
